adjust dir structure
