+++ src/mito2/src/manifest.rs
@@ -18,5 +18,6 @@ pub mod action;
 pub mod helper;
 #[allow(unused_variables)]
 pub mod manager;
-pub mod options;
 pub mod storage;
+#[cfg(test)]
+mod tests;
