+++ src/mito2/src/manifest/tests.rs
@@ -1,0 +1,1 @@
+mod checkpoint;
