+++ src/mito2/src/test_util.rs
@@ -30,8 +30,7 @@ use store_api::storage::RegionId;
 use crate::config::MitoConfig;
 use crate::engine::MitoEngine;
 use crate::error::Result;
-use crate::manifest::manager::RegionManifestManager;
-use crate::manifest::options::RegionManifestOptions;
+use crate::manifest::manager::{RegionManifestManager, RegionManifestOptions};
 use crate::metadata::{ColumnMetadata, RegionMetadataRef, SemanticType};
 use crate::request::{CreateRequest, RegionOptions};
 use crate::worker::WorkerGroup;
